fix(di) : di 관련 클래스 추가 및 연결
