+++ app/src/main/java/com/rsupport/mobile1/test/activity/data/WebHelper.kt
@@ -1,0 +1,5 @@
+package com.rsupport.mobile1.test.activity.data
+
+interface WebHelper {
+    suspend fun requestGettyImage(): List<GettyImage>?
+}
