+++ app/src/main/java/com/rsupport/mobile1/test/activity/domain/GettyImageRepositoryImpl.kt
@@ -1,10 +1,9 @@
 package com.rsupport.mobile1.test.activity.domain
 
-import com.rsupport.mobile1.test.activity.data.GettyImageDataSourceImpl
+import com.rsupport.mobile1.test.activity.data.GettyImageDataSource
+import javax.inject.Inject
 
-class GettyImageRepositoryImpl(private val web: GettyImageDataSourceImpl): GettyImageRepository {
-
-    override suspend fun getImage(): List<GettyImageResponse>? {
-        return web.requestGettyImage()?.map { it.mapper() }
-    }
+class GettyImageRepositoryImpl @Inject constructor(private val dataSource: GettyImageDataSource): GettyImageRepository {
+    override suspend fun getImage(): List<GettyImageResponse>? =
+        dataSource.requestGettyImage()?.map { it.mapper() }
 }
