+++ app/src/main/java/com/rsupport/mobile1/test/activity/presenter/MainActivity.kt
@@ -2,25 +2,21 @@ package com.rsupport.mobile1.test.activity.presenter
 
 import android.os.Bundle
 import android.util.Log
+import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
-import androidx.lifecycle.ViewModelProvider
 import com.rsupport.mobile1.test.R
-import com.rsupport.mobile1.test.activity.data.GettyImageDataSourceImpl
-import com.rsupport.mobile1.test.activity.domain.GettyImageRepositoryImpl
 import com.rsupport.mobile1.test.databinding.ActivityMainBinding
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class MainActivity : AppCompatActivity() {
 
     companion object {
         const val TAG = "GETTY_IMAGE"
     }
 
-    private val viewModel: MainViewModel by lazy {
-        val repository = GettyImageRepositoryImpl(GettyImageDataSourceImpl())
-        val factory = MainViewModel.Factory(repository)
-        ViewModelProvider(this, factory).get(MainViewModel::class.java)
-    }
+    private val viewModel: MainViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
