+++ app/src/main/java/com/rsupport/mobile1/test/activity/presenter/MainViewModel.kt
@@ -4,16 +4,18 @@ import android.util.Log
 import androidx.lifecycle.*
 import com.rsupport.mobile1.test.activity.domain.GettyImageRepositoryImpl
 import com.rsupport.mobile1.test.activity.domain.GettyImageResponse
+import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.*
+import javax.inject.Inject
 import kotlin.coroutines.CoroutineContext
 
-class MainViewModel(private val gettyImageRepository: GettyImageRepositoryImpl) : ViewModel(), CoroutineScope {
-
+@HiltViewModel
+class MainViewModel @Inject constructor(private val gettyImageRepository: GettyImageRepositoryImpl) : ViewModel(), CoroutineScope {
     override val coroutineContext: CoroutineContext
         get() = viewModelScope.coroutineContext
 
-    private val _gettyImageList = MutableLiveData<List<GettyImageResponse>>()
-    val gettyImageList: LiveData<List<GettyImageResponse>>
+    private val _gettyImageList = MutableLiveData<List<GettyImageResponse>?>()
+    val gettyImageList: LiveData<List<GettyImageResponse>?>
         get() = _gettyImageList
 
     fun requestImage() = launch (coroutineContext + Dispatchers.IO + CoroutineExceptionHandler { coroutineContext, throwable ->
@@ -27,12 +29,4 @@ class MainViewModel(private val gettyImageRepository: GettyImageRepositoryImpl) 
             Log.e(MainActivity.TAG, "result data is empty")
         }
     }
-
-
-    class Factory(private val gettyImageRepository: GettyImageRepositoryImpl) : ViewModelProvider.Factory {
-        override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            return MainViewModel(gettyImageRepository) as T
-        }
-    }
-
 }
